net/nfp: move shared target logic to own source file
Move the needed logic of nfp_target.h and nfp_cppat.h
to a new source file nfp_target.c, and delete the logic
which are not used at all.

The code is moved verbatim from the header file to the source
file, no functional change.

This mainly to mimic the source file structure in the kernel
driver, and also gives a nice decrease in driver size.

Signed-off-by: Chaoyong He <chaoyong.he@corigine.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund@corigine.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>