reorder: introduce API to obtain memory footprint
At present, it is not easy to determine the memory requirement for the
reorder buffer without delving into its implementation details.
To facilitate the use of reorder buffer with user allocation, a new API
`rte_reorder_memory_footprint_get()` is introduced.
This API will provide the amount of required memory for reorder buffer.

Signed-off-by: Volodymyr Fialko <vfialko@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>