devtools: allow patch to multiple groups for the same driver
The PMD's source code resides in the ./drivers folder of the
DPDK project and split into the several groups depending on the
PMD class (common, net, regex, etc.).

For some vendors the drivers of different classes operate over
the same hardware, for example Nvidia PMDs operate over ConnectX
NIC series. It often happens the same minor fixes should be applied
to the multiple drivers of the same vendor in the different classes.

The check-git-log.sh script checks the consistence of patch
affected files and patch commit message headline and prevents
updating multiple drivers in single commit.

This patch mitigates this strict check and allows to update
multiple drivers in different classes for the single vendor.

Signed-off-by: Viacheslav Ovsiienko <viacheslavo@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>