net/hns3: simplify for hardware csum offloading
If the NIC support simple BD mode, the hardware will calculate
the checksum from the start position of checksum and fill the
checksum result to the offset position, which simple the
HW operations of calculating the type and header length of
L3/L4. Add this mode for hns3 pmd when the packet type
is L4.

Signed-off-by: Dongdong Liu <liudongdong3@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>