mldev: remove xstats APIs from library
This change is a preparatoy step for revised xstats APIs.
Revised xstats APIs support reporting device and per model
stats, which is based on eventdev xstats.

Removed xstats APIs from spec and library implementation.
Disabled reporting xstats in test application and disabled
xstats functions in drivers. Renamed stats_get function to
throughput_get.

This change is needed as the revised APIs are not backward
compatible with the current xstats.

Signed-off-by: Srikanth Yalavarthi <syalavarthi@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>