net/vmxnet3: add support for capability registers
This patch enhances vmxnet3 to support capability registers which
allows it to enable features selectively. The DCR register tracks
the capabilities vmxnet3 device supports. The PTCR register states
the capabilities that the passthrough device supports.

With the help of these registers, vmxnet3 can enable only those
features which the passthrough device supports. This allows
smooth transition to Uniform-Passthrough (UPT) mode if the virtual
NIC requests it. If PTCR register returns nothing or error it means
UPT is not being requested and vnic will continue in emulation mode.

Signed-off-by: Ronak Doshi <doshir@vmware.com>
Acked-by: Jochen Behrens <jbehrens@vmware.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>