net/ixgbe: consider DCB/VMDq conf when getting RSS conf
The mrqe field of MRQC register is an enum. From the Intel 82599 datasheet,
we know that these values below for the mrqe field are all related to RSS
configuration:
	0000b = RSS disabled.
	0001b = RSS only -- Single set of RSS 16 queues.
	0010b = DCB enabled and RSS disabled -- 8 TCs, each allocated 1 queue.
 	0011b = DCB enabled and RSS disabled -- 4 TCs, each allocated 1 queue.
	0100b = DCB and RSS -- 8 TCs, each allocated 16 RSS queues.
	0101b = DCB and RSS -- 4 TCs, each allocated 16 RSS queues.
	1000b = Virtualization only -- 64 pools, no RSS, each pool allocated
		2 queues.
	1010b = Virtualization and RSS -- 32 pools, each allocated 4 RSS queues.
	1011b = Virtualization and RSS -- 64 pools, each allocated 2 RSS queues.

The ixgbe pmd will check whether the rss is enabled or not when getting
rss conf. So, beside comparing the value of mrqe field with xxx0b and
xxx1b, we also need to consider the other configurations, such as
DCB + RSS or VMDQ + RSS. Otherwise, we may not get the correct rss conf
in some cases, such as when we use DCB and RSS with 8 TCs which corresponds
to 0100b for the mrqe field.

Signed-off-by: Min Zhou <zhoumin@loongson.cn>
Signed-off-by: 0-day Robot <robot@bytheb.org>