app/pdump: make dpdk-pdump can run multiple simultaneously
The pdump doesn't support start the multiple separate process for different
queues or ports to monitor the packets, because it apply the same name for
memory space. This commit will use the device_id, port and queue to name
the memory space and it will support this function.

Adjust the value of RTE_MEMZONE_NAMESIZE to adapt the name length.

Signed-off-by: Peng Zhang <peng.zhang@corigine.com>
Reviewed-by: Chaoyong He <chaoyong.he@corigine.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund@corigine.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>