net/ice/base: fix incorrect defines for DCBx
In all cases but one, CEE mode is defined as 0x01 and IEEE
mode is defined as 0x02.  in past these values were swapped.
This is causing the DCB information sent from the FW agent to
be parsed with incorrect structures and resulting in incorrect values.

Change the defines to match what they are in the rest of the kernel.

Fixes: 97e32e8d4870 ("net/ice/base: complete pending LLDP MIB")
Signed-off-by: Dave Ertman <david.m.ertman@intel.com>
Signed-off-by: Qiming Yang <qiming.yang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>